requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via aiohttp
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
faiss-cpu==1.13.0
    # via agent (pyproject.toml)
fastapi==0.121.2
    # via agent (pyproject.toml)
filelock==3.20.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via agent (pyproject.toml)
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   openai
httpx-sse==0.4.3
    # via langchain-community
huggingface-hub==0.36.0
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via torch
jiter==0.12.0
    # via openai
joblib==1.5.2
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.0.7
    # via agent (pyproject.toml)
langchain-classic==1.0.0
    # via langchain-community
langchain-community==0.4.1
    # via agent (pyproject.toml)
langchain-core==1.0.5
    # via
    #   langchain
    #   langchain-classic
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==1.0.3
    # via agent (pyproject.toml)
langchain-text-splitters==1.0.0
    # via langchain-classic
langgraph==1.0.3
    # via langchain
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.4
    # via langgraph
langgraph-sdk==0.2.9
    # via langgraph
langsmith==0.4.43
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
markupsafe==3.0.3
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
mpmath==1.3.0
    # via sympy
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
networkx==3.5
    # via torch
numpy==2.3.5
    # via
    #   faiss-cpu
    #   langchain-community
    #   scikit-learn
    #   scipy
    #   transformers
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.5
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvshmem-cu12==3.3.20
    # via torch
nvidia-nvtx-cu12==12.8.90
    # via torch
openai==2.8.0
    # via
    #   agent (pyproject.toml)
    #   langchain-openai
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   faiss-cpu
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   transformers
pillow==12.0.0
    # via sentence-transformers
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.12.4
    # via
    #   agent (pyproject.toml)
    #   fastapi
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via langchain-community
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   langchain-classic
    #   langchain-community
    #   langchain-core
    #   transformers
    #   uvicorn
regex==2025.11.3
    # via
    #   tiktoken
    #   transformers
requests==2.32.5
    # via
    #   huggingface-hub
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
safetensors==0.6.2
    # via transformers
scikit-learn==1.7.2
    # via sentence-transformers
scipy==1.16.3
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.1.2
    # via agent (pyproject.toml)
setuptools==80.9.0
    # via torch
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.44
    # via
    #   langchain-classic
    #   langchain-community
starlette==0.49.3
    # via fastapi
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.12.0
    # via langchain-openai
tokenizers==0.22.1
    # via transformers
torch==2.9.1
    # via
    #   agent (pyproject.toml)
    #   sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.57.1
    # via sentence-transformers
triton==3.5.1
    # via torch
typing-extensions==4.15.0
    # via
    #   fastapi
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   sqlalchemy
    #   torch
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via agent (pyproject.toml)
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
